tiny changes with MainMenu

+++ src/th06_102h.exe/Global/FUN_0042a570.c
@@ -11,9 +11,9 @@ undefined4 FUN_0042a570(ReplayManager *param_1)
     g_CurFrameInput = g_CurFrameInput & 0xfe08 | param_1->replay_inputs->input_key;
     g_IsEigthFrameOfHeldInput = 0;
     if (g_LastFrameInput == g_CurFrameInput) {
-      if ((0x1d < g_NumOfFramesInputsWereHeld) &&
-         (g_IsEigthFrameOfHeldInput = (ushort)((g_NumOfFramesInputsWereHeld & 7) == 0),
-         0x25 < g_NumOfFramesInputsWereHeld)) {
+      if ((0x1d < (ushort)g_NumOfFramesInputsWereHeld) &&
+         (g_IsEigthFrameOfHeldInput = (short)((g_NumOfFramesInputsWereHeld & 7U) == 0),
+         0x25 < (ushort)g_NumOfFramesInputsWereHeld)) {
         g_NumOfFramesInputsWereHeld = 0x1e;
       }
       g_NumOfFramesInputsWereHeld = g_NumOfFramesInputsWereHeld + 1;
